requirements.txt
numpy
playsound
pynput
rapidfuzz
pyaudio
openai-whisper